:zap: perf(PlanReviewTests.py): Refactored to modularize according to test levels
Created a more standardized way of making tests. Essentially each test in now called
with the following syntax:
Test(pd,**kwargs) where pd is the named tuple used for storing RS objects
and kwargs is the optional addition arguments.

Several separate files were created for:
User definitions: ReviewDefinitions.py
Beamset checks: BeamSetReviewTests.py
Plan checks: PlanReviewTests.py
Exam checks: ExamTests.py

Icons were moved.
