+++ library/PlanReview/ReviewDefinitions.py
@@ -1,0 +1,108 @@
+import os
+
+icon_dir = os.path.join(os.path.dirname(__file__), ".\Icons")
+RED_CIRCLE = os.path.join(icon_dir, "red_circle_icon.png")
+GREEN_CIRCLE = os.path.join(icon_dir, "green_circle_icon.png")
+BLUE_CIRCLE = os.path.join(icon_dir, "blue_circle_icon.png")
+
+#
+# LOG PARSING INFO
+LOG_DIR = r"Q:\\RadOnc\RayStation\RayScripts\logs"
+DEV_LOG_DIR = r"Q:\\RadOnc\RayStation\RayScripts\dev_logs"
+KEEP_PHRASES = [("Critical", "CRITICAL"), ("Warnings", "WARNING"),
+                ("Info", "INFO"), ("Debug", "DEBUG")]
+#
+# EXAM DEFAULTS
+# TIME ELAPSED BETWEEN PLAN AND CT
+DAYS_SINCE_SIM = 14
+#
+# CONTOURING DEFAULTS
+BOLUS_NAMES = ["bolus"]
+# def - check the front edges of the couch and suspended headboard
+NO_FLY_NAME = "NoFlyZone_PRV"
+#
+# PLANNING DEFAULTS
+DOSE_FRACTION_PAIRS = [(4, 2000), (5, 2000)]  # Often mixed up fractionations
+#
+EDW_MU_LIMIT = 20.
+#
+# DOSE TOLERANCES
+NO_FLY_DOSE = 100.  # cGy
+PACEMAKER_DOSE = 200.  # cGy
+#
+# DOSE GRID PREFERENCES
+DOSE_GRID_DEFAULT = 0.2  # 2 mm
+
+GRID_PREFERENCES = {
+    'SBRT': {
+        'PLAN_NAMES': ['LUL', 'LLL', 'RUL', 'RML', 'RLL',
+                       'LuLU_SBR', 'LuLL_SBR', 'LuRU_SBR', 'LuRM_SBR', 'LuRL_SBR',
+                       'BreR_SBR', 'BreL_SBR', 'Abdo_SBR', 'LivR_SBR', 'Panc_SBR',
+                       'Pelv_SBR', 'HipR_SBR', 'HipL_SBR'],
+        'DOSE_GRID': 0.15,  # 1.5 mm
+        'FRACTION_SIZE_LIMIT': 801,  # cGy
+        'SLICE_THICKNESS': 0.2,  # 2.0 mm
+    },
+    'SBRT_FINE': {
+        'PLAN_NAMES': ['Pros_SBR', 'Brai_SBR', 'NecB_SBR', 'NecL_SBR', 'SpiT_SBR', 'SpiL_SBR', 'SpiC_SBR'],
+        'DOSE_GRID': 0.15,  # 1.5 mm
+        'FRACTION_SIZE_LIMIT': 801,  # cGy
+        'SLICE_THICKNESS': 0.1,  # 2.0 mm
+    },
+    'SRS': {
+        'PLAN_NAMES': ['SRS', 'FSR'],
+        'DOSE_GRID': 0.1,  # 1.0 mm
+        'FRACTION_SIZE_LIMIT': 1500,  # cGy
+        'SLICE_THICKNESS': 0.1,  # 1.0 mm
+    },
+    'TBI': {
+        'PLAN_NAMES': ['TBI'],
+        'DOSE_GRID': 0.5,  # 5 mm
+        'FRACTION_SIZE_LIMIT': None,  # Don't check
+        'SLICE_THICKNESS': 0.4,  # 4 mm
+    },
+    'VMAT': {
+        'PLAN_NAMES': ['VMA', '3CA'],
+        'DOSE_GRID': 0.3,  # 3 mm
+        'FRACTION_SIZE_LIMIT': 800,  # cGy
+        'SLICE_THICKNESS': 0.3,  # 3 mm
+    },
+    'THI': {
+        'PLAN_NAMES': ['THI', 'T3D'],
+        'DOSE_GRID': 0.3,  # 3 mm
+        'FRACTION_SIZE_LIMIT': 800,  # cGy
+        'SLICE_THICKNESS': 0.3,  # 3 mm
+    },
+    '3D': {
+        'PLAN_NAMES': ['3DC'],
+        'DOSE_GRID': 0.4,  # 3 mm
+        'FRACTION_SIZE_LIMIT': 800,  # cGy
+        'SLICE_THICKNESS': 0.4,  # 4 mm
+    },
+}
+#
+# FIELD OF VIEW SETTINGS
+FIELD_OF_VIEW_PREFERENCES = {'NAME': 'FOV_Reconstructed',
+                             'WALL_SUFFIX': '_Wall',
+                             'CONTRACTION': 0.5,  # cm
+                             'NAME_INTERSECTION': 'FOV_EXT_INTERSECT',
+                             'SI_PTV_BUFFER': 2.0,  # cm
+                             }
+
+MCS_TOLERANCES = {'MCS': {'MEAN': 0.369,
+                          'SIGMA': 0.152},
+                  'LSV': {'MEAN': 0.694,
+                          'SIGMA': 0.134},
+                  'AAV': {'MEAN': 0.522,
+                          'SIGMA': 0.188},
+                  }
+
+TOMO_DATA = {'MACHINES': ['HDA0488', 'HDA0477'],
+             'PLAN_TR_SUFFIX': r'_Tr',
+             'LATERAL_ISO_MARGIN': 2.,  # cm
+             'SUPPORTS': ['TomoCouch']
+             }
+
+TRUEBEAM_DATA = {'MACHINES': ['TrueBeam', 'TrueBeamSTx'],
+                 'SUPPORTS': ['TrueBeamCouch', 'CivcoBaseShell_Cork', 'CivcoInclineShell_Wax',
+                              'Sframe_F1_TBCouch_HN']}
